+++ src/lib/pages/custom-network/components/FailedAddCustomMinitiaModal.tsx
@@ -42,7 +42,7 @@ export const FailedAddCustomMinitiaModal = ({
       </ModalHeader>
       <ModalBody overflow="overlay" pt={2}>
         <Text variant="body2" color="text.dark" textAlign="center">
-          There is something wrong with your JSON file. So that the Minitia is
+          There is something wrong with your JSON file. So that the Rollup is
           cannot be added to InitiaScan.
         </Text>
       </ModalBody>