+++ src/lib/pages/custom-network/manual/index.tsx
@@ -137,14 +137,14 @@ export const AddNetworkManual = () => {
   };
 
   const handleActionLabel = () => {
-    if (currentStepIndex === 2) return "Save new Minitia";
+    if (currentStepIndex === 2) return "Save new Rollup";
 
     return "Next";
   };
 
   return (
     <>
-      <CelatoneSeo pageName="Add Minitias" />
+      <CelatoneSeo pageName="Add Rollups" />
       <Flex position="sticky" top={0} left={0} w="full" zIndex={2}>
         <AddNetworkStepper currentStepIndex={currentStepIndex} />
       </Flex>
@@ -174,7 +174,7 @@ export const AddNetworkManual = () => {
           ) : undefined,
         }}
         actionLabel={handleActionLabel()}
-        helperText="The added custom Minitia on Initiascan will be stored locally on your device."
+        helperText="The added custom Rollup on Initiascan will be stored locally on your device."
         sx={{
           backgroundColor: "background.main",
           borderColor: "gray.700",